moved validateProject into middleware file
